pip_requirements.txt
h5py==3.10.0 # 3.11
lpips==0.1.4 # 0.1.4
matplotlib==3.5.1 # 3.7.5
monai==0.8.1 # 1.3.2
nibabel==5.1.0
numba==0.55.1 # 0.58.1
numpy==1.21.5
opencv_python==4.8.1.78 # 4.10.0
Pillow==10.1.0
pydicom==2.4.3
pytorch-lightning==1.6.3 # 1.9 
pytorch_msssim==0.2.1 # 1.0.0
torchinfo==1.8.0 # 1.8.0
torchio==0.19 # 0.19.9
torchmetrics==0.8.2
PyYAML==6.0
setuptools==59.5.0
scikit_image==0.19.2 # 0.21.0
scikit_learn==1.0.2 # 1.3.2
scipy==1.8.0
seaborn==0.11.2 # 0.13.2
tqdm==4.62.3
# xformers==0.0.22.post7
xformers==0.0.10 # 0.0.10
plotly==5.8.0 # 5.23.0
wandb==0.16.1 # wandb-0.17.5

# python 3.8 
# torch 1.12
